+++ src/store/user/user.initialState.js
@@ -1,0 +1,9 @@
+export const initialState = {
+	isAuth: false,
+	name: '',
+	email: '',
+	token: '',
+	error: '',
+	registerError: '',
+	registerSuccess: false,
+};
