+++ src/store/user/user.js
@@ -1,48 +1,6 @@
-import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
-import { apiService } from '../../sesrvice';
-
-const initialState = {
-	isAuth: false,
-	name: '',
-	email: '',
-	token: '',
-	error: '',
-	registerError: '',
-	registerSuccess: false,
-};
-
-const fetchLogin = createAsyncThunk(
-	'courses/fetchLogin',
-	async (body, { rejectWithValue }) => {
-		try {
-			const response = await apiService.login('login', body);
-			const result = await response.json();
-			if (!response.ok) {
-				throw new Error(`${result.errors[0]}`);
-			}
-			return result;
-		} catch (error) {
-			return rejectWithValue(error.message);
-		}
-	}
-);
-
-const fetchRegistration = createAsyncThunk(
-	'courses/fetchRegistration',
-	async (body, { rejectWithValue }) => {
-		try {
-			const response = await apiService.registration('register', body);
-			const result = await response.json();
-			console.log(result);
-			if (!response.ok) {
-				throw new Error(`${result.errors[0]}`);
-			}
-			return result;
-		} catch (error) {
-			return rejectWithValue(error.message);
-		}
-	}
-);
+import { createSlice } from '@reduxjs/toolkit';
+import { initialState } from './user.initialState';
+import { fetchLogin, fetchRegistration } from './actions.user';
 
 const userSlice = createSlice({
 	name: 'user',
@@ -95,11 +53,4 @@ const getRegistrationData = (state) => ({
 const { logout } = userSlice.actions;
 const userReducer = userSlice.reducer;
 
-export {
-	getAuthData,
-	getRegistrationData,
-	fetchRegistration,
-	fetchLogin,
-	logout,
-	userReducer,
-};
+export { getAuthData, getRegistrationData, logout, userReducer };
